+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * Use SEND_WITH_CALLBACK in NettyHttpServiceClient if the request was
   configured for connectionSharing and SSL
 
+* Add QueryOption.SELECT_LINKS for selecting fields marked with
+  PropertyUsageOption.LINK and including the link values in the query results.
+  The link selection will be used for future graph query support and
+  automatic link content expansion.
+
 * Add transaction flow support, across related operations, similar to
   authorization context and context id flow support